go-hass: initial commit
This commit puts in place the basic infrastructure for a HASS client
in Go, with Get request support, a basic testing harness, and a demo
CLI for the library.

This is a thin client kind of library: there is not much logic to
apply in a HASS API client, at least for now. Since this is a
playground type of project, this may change, and we may decide to add
more client-side smarts (e.g. in caching) at a future time.