Add basic support for /services API calls.
Unfortunately /services is poorly documented, the documentation is out
of date and does not cover the logic of the complex fields that make
this part of the API really hard to grok. So this is just a starting
point for now, until upstream fixes this.

Also a minor API change: API types are now public (they are meant to
be manipulated outside this package).